requirements_LOCAL_1213.txt
streamlit
pandas
streamlit_folium
folium
numpy
datetime
openmeteo_requests
requests_cache
retry_requests
plotly
scipy
streamlit-cookies-controller
dotmap
rws-ddlpy
